Restructure repo's file structure
To cope with Markdown (at least in GOGS) not being able to start ordered
lists with 0, we incremented each chapter number, so that the source
match the rendered README file.

Moreover, we have decided not to hide the .svg files in folder anymore,
but to move them at the root of the repo.

Signed-off-by: Étienne Boisseau-Sierra <etienne.boisseau-sierra@unipart.io>